fix(pricing): move modelsdev cache to backend
